# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "spelunk-server",
    "description": "Shared memory server for spelunk. Stores decisions, requirements, and context for a team and serves them over HTTP.",
    "contact": {
      "name": "spelunk",
      "url": "https://github.com/spelunk-cloud/spelunk"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/projects/{project_id}/explore": {
      "post": {
        "tags": [
          "inference"
        ],
        "summary": "Run an LLM reasoning loop over caller-supplied context chunks.\nThe CLI retrieves relevant chunks from its local index and sends them alongside\nthe question. **The server does not store context chunks.**",
        "description": "Returns an SSE stream with events: `thought`, `answer`, `done`, `error`.\nReturns 503 if no LLM is configured.",
        "operationId": "explore",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "Project slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExploreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "SSE stream: thought/answer/done/error events"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "No LLM configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/v1/projects/{project_id}/index/embed": {
      "post": {
        "tags": [
          "index"
        ],
        "summary": "Generate embeddings for code chunks. The server encodes each chunk and returns the\nvectors. **The server does not store the vectors**: the CLI is the only persistent\nstore for index data.",
        "description": "Returns 400 if no embedder is configured.\nReturns 413 if the batch exceeds 256 chunks.\nReturns 429 (with `Retry-After`) if the embed admission queue is full.",
        "operationId": "index_embed",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "Project slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmbedRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Embedding vectors as raw little-endian f32 bytes, row-major [n_chunks x dim] in request order (not stored server-side)",
            "content": {
              "application/octet-stream": {}
            }
          },
          "400": {
            "description": "No embedder configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "413": {
            "description": "Batch exceeds 256 chunks",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "Embed admission queue full; retry after the given delay",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EmbedChunkIn": {
        "type": "object",
        "description": "A single chunk to embed.",
        "required": [
          "chunk_id",
          "content"
        ],
        "properties": {
          "chunk_id": {
            "type": "string",
            "description": "Opaque CLI-assigned identifier (e.g. blake3 hash of file + offset). Echoed back verbatim."
          },
          "content": {
            "type": "string",
            "description": "Raw text content to embed."
          }
        }
      },
...
      "EmbedRequest": {
        "type": "object",
        "description": "Request body for `POST /v1/projects/{project_id}/index/embed`.",
        "required": [
          "chunks"
        ],
        "properties": {
          "chunks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EmbedChunkIn"
            },
            "description": "Chunks to embed. Maximum 256 per request."
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "Consistent JSON error body: `{\"error\": {\"code\": \"...\", \"message\": \"...\"}}`.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
      "ErrorDetail": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ExploreContextChunk": {
        "type": "object",
        "description": "A single context chunk supplied by the CLI for `/explore`.",
        "required": [
          "file",
          "start_line",
          "end_line",
          "content"
        ],
        "properties": {
          "content": {
            "type": "string"
          },
          "end_line": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "file": {
            "type": "string"
          },
          "start_line": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
      "ExploreRequest": {
        "type": "object",
        "description": "Request body for `POST /v1/projects/{project_id}/explore`.",
        "required": [
          "question"
        ],
        "properties": {
          "context_chunks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExploreContextChunk"
            }
          },
          "max_turns": {
            "type": "integer",
            "minimum": 0
          },
          "question": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API key",
        "description": "Pass as `Authorization: Bearer <key>`. Not required when no key is configured on the server."
      }
    }
  },
  "security": [
    {
      "bearer_auth": []
    }
  ],
...
}